🛠 EdgeToEdgeBottomSheetDialog common widget 으로 변경
