+++ app/src/main/java/com/mashup/ui/main/popup/MainBottomPopup.kt
@@ -33,7 +33,6 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import androidx.core.os.bundleOf
 import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
@@ -45,6 +44,7 @@ import com.mashup.constant.log.LOG_COMMON_POPUP_CANCEL
 import com.mashup.constant.log.LOG_COMMON_POPUP_CONFIRM
 import com.mashup.core.common.utils.getDrawableResIdByName
 import com.mashup.core.common.utils.keyboard.RootViewDeferringInsetsCallback
+import com.mashup.core.common.widget.EdgeToEdgeBottomSheetDialog
 import com.mashup.core.ui.colors.Gray500
 import com.mashup.core.ui.colors.Gray950
 import com.mashup.core.ui.colors.White
@@ -88,21 +88,10 @@ class MainBottomPopup : BottomSheetDialogFragment() {
         }
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
-        object : BottomSheetDialog(requireContext(), theme) {
-            override fun onAttachedToWindow() {
-                super.onAttachedToWindow()
-
-                window?.let {
-                    WindowCompat.setDecorFitsSystemWindows(it, false)
-                }
-
-                findViewById<View>(com.google.android.material.R.id.container)?.apply {
-                    fitsSystemWindows = false
-                }
-
-                findViewById<View>(com.google.android.material.R.id.coordinator)?.fitsSystemWindows = false
-            }
-        }
+        EdgeToEdgeBottomSheetDialog(
+            context = requireContext(),
+            theme = theme
+        )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
